# Improve log perfomance