--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,7 @@ Most recent version is listed first.
 - ong/log: Conform to log/slog: https://github.com/komuw/ong/pull/357
            Eliminate duplicate logIDs
 - ong/log: Integrate ong packages with new logger: https://github.com/komuw/ong/pull/358
+- ong/log: Improve log perfomance: https://github.com/komuw/ong/pull/359
 
 # v0.0.73
 - Remove /ex/exp: https://github.com/komuw/ong/pull/352